refactor(ProForma): Consolidate parser, data, writer, tokenizer, error into single files
Combine all ProForma-related files into a single header (ProForma.h) and
implementation (ProForma.cpp) to simplify the API and reduce file count.

Changes:
- Create ProForma.h with all public API (enums, data structures, ProFormaParser class)
- Create ProForma.cpp with internal detail classes (ProFormaTokenizer, ProFormaWriter,
  ProFormaParserImpl) and all public method implementations
- Update ProFormaDataJson.h to include ProForma.h
- Update all files that included old ProForma headers
- Update CMake sources.cmake files
- Update pyOpenMS pxd wrappers to use new header path
- Remove tokenizer tests from test file (now internal implementation detail)

Removed files:
- ProFormaData.h, ProFormaError.h, ProFormaParser.h, ProFormaTokenizer.h, ProFormaWriter.h
- ProFormaError.cpp, ProFormaParser.cpp, ProFormaTokenizer.cpp, ProFormaWriter.cpp
