+++ src/openms/include/OpenMS/CHEMISTRY/sources.cmake
@@ -24,12 +24,8 @@ ModificationsDB.h
 MonosaccharideDB.h
 NASequence.h
 NucleicAcidSpectrumGenerator.h
-ProFormaData.h
+ProForma.h
 ProFormaDataJson.h
-ProFormaError.h
-ProFormaParser.h
-ProFormaTokenizer.h
-ProFormaWriter.h
 ProteaseDB.h
 ProteaseDigestion.h
 Residue.h
